ещё методы

+++ DirectXFold/DirectXFold/inc/SphericalMeshLoader.h
@@ -9,7 +9,7 @@ public:
 private:
 	static std::vector<std::string> parseString(std::string str, char delimiter);
 	static std::vector<double> getDoubleValues(std::vector<std::string> strings);
-	static std::vector<double> getIntValues(std::vector<std::string> strings);
+	static std::vector<int> getIntValues(std::vector<std::string> strings);
 
 };
